Refactor Transaction printing methods into a separate package.
